[MIK-1391] 【Design】オファー機能のデザイン修正Sửa design chức năng offer

+++ resources/views/admin/offers/detail.blade.php
@@ -51,11 +51,11 @@
             <div class="col-sm-4">
               {{ Carbon\Carbon::parse($offer->start_time_from)->format('H:i') }} &nbsp ~ &nbsp
               @php
-                $startHour = (int)Carbon\Carbon::parse($offer->start_time_from)->format('H');
-                $endHour = (int)Carbon\Carbon::parse($offer->start_time_to)->format('H');
-                $endMinute = (int)Carbon\Carbon::parse($offer->start_time_to)->format('i');
+                $startHour = Carbon\Carbon::parse($offer->start_time_from)->format('H');
+                $endHour = Carbon\Carbon::parse($offer->start_time_to)->format('H');
+                $endMinute = Carbon\Carbon::parse($offer->start_time_to)->format('i');
 
-                if ($endHour < $startHour) {
+                if ((int)$endHour < (int)$startHour) {
                   switch ($endHour) {
                   case 0:
                   $endHour = 24;
